* Incorperate builder methods into GenericArray directly. * Use ManuallyDrop now that it's stable
