+++ Cargo.toml
@@ -15,10 +15,6 @@ name = "generic_array"
 typenum = "1.9"
 serde = { version = "1.0", optional = true, default-features = false }
 
-[dependencies.nodrop]
-version = "~0.1"
-default-features = false
-
 [dev_dependencies]
 # this can't yet be made optional, see https://github.com/rust-lang/cargo/issues/1596
 serde_json = "1.0"
