+++ src/lib.rs
@@ -1,5 +1,6 @@
 //! This crate implements a structure that can be used as a generic array type.use
-//! Core Rust array types `[T; N]` can't be used generically with respect to `N`, so for example this:
+//! Core Rust array types `[T; N]` can't be used generically with
+//! respect to `N`, so for example this:
 //!
 //! ```{should_fail}
 //! struct Foo<T, N> {
@@ -9,7 +10,8 @@
 //!
 //! won't work.
 //!
-//! **generic-array** exports a `GenericArray<T,N>` type, which lets the above be implemented as:
+//! **generic-array** exports a `GenericArray<T,N>` type, which lets
+//! the above be implemented as:
 //!
 //! ```
 //! # use generic_array::{ArrayLength, GenericArray};
@@ -18,7 +20,9 @@
 //! }
 //! ```
 //!
-//! The `ArrayLength<T>` trait is implemented by default for [unsigned integer types](../typenum/uint/index.html) from [typenum](../typenum/index.html).
+//! The `ArrayLength<T>` trait is implemented by default for
+//! [unsigned integer types](../typenum/uint/index.html) from
+//! [typenum](../typenum/index.html).
 //!
 //! For ease of use, an `arr!` macro is provided - example below:
 //!
@@ -32,31 +36,34 @@
 //! # }
 //! ```
 
-#![deny(missing_docs)]
+//#![deny(missing_docs)]
 #![no_std]
+
 pub extern crate typenum;
-extern crate nodrop;
+
 #[cfg(feature = "serde")]
 extern crate serde;
-pub mod arr;
-pub mod iter;
-pub use iter::GenericArrayIter;
+
 mod hex;
 mod impls;
 
 #[cfg(feature = "serde")]
 pub mod impl_serde;
 
-pub mod builder;
-
-pub use self::builder::{IterableArrayLength, RecursiveArrayBuilder};
+use core::{mem, ptr, slice};
 
 use core::marker::PhantomData;
+use core::mem::ManuallyDrop;
 pub use core::mem::transmute;
 use core::ops::{Deref, DerefMut};
-use core::slice;
+
 use typenum::bit::{B0, B1};
 use typenum::uint::{UInt, UTerm, Unsigned};
+
+#[cfg_attr(test, macro_use)]
+pub mod arr;
+pub mod iter;
+pub use iter::GenericArrayIter;
 
 /// Trait making `GenericArray` work, marking types to be used as length of an array
 pub unsafe trait ArrayLength<T>: Unsigned {
@@ -144,6 +151,62 @@ where
     }
 }
 
+struct ArrayBuilder<T, N: ArrayLength<T>> {
+    array: ManuallyDrop<GenericArray<T, N>>,
+    position: usize,
+}
+
+impl<T, N: ArrayLength<T>> ArrayBuilder<T, N> {
+    fn new() -> ArrayBuilder<T, N> {
+        ArrayBuilder {
+            array: ManuallyDrop::new(unsafe { mem::uninitialized() }),
+            position: 0,
+        }
+    }
+
+    fn into_inner(self) -> GenericArray<T, N> {
+        let array = unsafe { ptr::read(&self.array) };
+
+        mem::forget(self);
+
+        ManuallyDrop::into_inner(array)
+    }
+}
+
+impl<T, N: ArrayLength<T>> Drop for ArrayBuilder<T, N> {
+    fn drop(&mut self) {
+        for value in self.array.iter_mut().take(self.position) {
+            unsafe {
+                ptr::drop_in_place(value);
+            }
+        }
+    }
+}
+
+struct ArrayConsumer<T, N: ArrayLength<T>> {
+    array: ManuallyDrop<GenericArray<T, N>>,
+    position: usize,
+}
+
+impl<T, N: ArrayLength<T>> ArrayConsumer<T, N> {
+    fn new(array: GenericArray<T, N>) -> ArrayConsumer<T, N> {
+        ArrayConsumer {
+            array: ManuallyDrop::new(array),
+            position: 0,
+        }
+    }
+}
+
+impl<T, N: ArrayLength<T>> Drop for ArrayConsumer<T, N> {
+    fn drop(&mut self) {
+        for i in self.position..N::to_usize() {
+            unsafe {
+                ptr::drop_in_place(self.array.get_unchecked_mut(i));
+            }
+        }
+    }
+}
+
 impl<T, N> GenericArray<T, N>
 where
     N: ArrayLength<T>,
@@ -152,12 +215,21 @@ where
     ///
     /// If the generator function panics while initializing the array,
     /// any already initialized elements will be dropped.
-    #[inline]
     pub fn generate<F>(f: F) -> GenericArray<T, N>
     where
         F: Fn(usize) -> T,
     {
-        builder::generate(f)
+        let mut destination = ArrayBuilder::new();
+
+        for (i, dst) in destination.array.iter_mut().enumerate() {
+            unsafe {
+                ptr::write(dst, f(i));
+            }
+
+            destination.position += 1;
+        }
+
+        destination.into_inner()
     }
 
     /// Map a function over a slice to a `GenericArray`.
@@ -167,20 +239,31 @@ where
     pub fn map_slice<S, F: Fn(&S) -> T>(s: &[S], f: F) -> GenericArray<T, N> {
         assert_eq!(s.len(), N::to_usize());
 
-        builder::generate(|i| f(unsafe { s.get_unchecked(i) }))
+        Self::generate(|i| f(unsafe { s.get_unchecked(i) }))
     }
 
     /// Maps a `GenericArray` to another `GenericArray`.
     ///
     /// If the mapping function panics, any already initialized elements in the new array
     /// will be dropped, AND any unused elements in the source array will also be dropped.
-    #[inline]
     pub fn map<U, F>(self, f: F) -> GenericArray<U, N>
     where
         F: Fn(T) -> U,
         N: ArrayLength<U>,
     {
-        builder::map(self, f)
+        let mut source = ArrayConsumer::new(self);
+        let mut destination = ArrayBuilder::new();
+
+        for (dst, src) in destination.array.iter_mut().zip(source.array.iter()) {
+            unsafe {
+                ptr::write(dst, f(ptr::read(src)));
+            }
+
+            source.position += 1;
+            destination.position += 1;
+        }
+
+        destination.into_inner()
     }
 
     /// Maps a `GenericArray` to another `GenericArray` by reference.
@@ -192,7 +275,7 @@ where
         F: Fn(&T) -> U,
         N: ArrayLength<U>,
     {
-        builder::generate(|i| f(unsafe { self.get_unchecked(i) }))
+        GenericArray::generate(|i| f(unsafe { self.get_unchecked(i) }))
     }
 
     /// Combines two `GenericArray` instances and iterates through both of them,
@@ -200,13 +283,31 @@ where
     ///
     /// If the mapping function panics, any already initialized elements in the new array
     /// will be dropped, AND any unused elements in the source arrays will also be dropped.
-    #[inline]
     pub fn zip<B, U, F>(self, rhs: GenericArray<B, N>, f: F) -> GenericArray<U, N>
     where
         F: Fn(T, B) -> U,
         N: ArrayLength<B> + ArrayLength<U>,
     {
-        builder::zip(self, rhs, f)
+        let mut left = ArrayConsumer::new(self);
+        let mut right = ArrayConsumer::new(rhs);
+
+        let mut destination = ArrayBuilder::new();
+
+        for (dst, (lhs, rhs)) in
+            destination.array.iter_mut().zip(left.array.iter().zip(
+                right.array.iter(),
+            ))
+        {
+            unsafe {
+                ptr::write(dst, f(ptr::read(lhs), ptr::read(rhs)));
+            }
+
+            destination.position += 1;
+            left.position += 1;
+            right.position += 1;
+        }
+
+        destination.into_inner()
     }
 
     /// Combines two `GenericArray` instances and iterates through both of them by reference,
@@ -218,11 +319,10 @@ where
         F: Fn(&T, &B) -> U,
         N: ArrayLength<B> + ArrayLength<U>,
     {
-        builder::generate(|i| unsafe {
+        GenericArray::generate(|i| unsafe {
             f(self.get_unchecked(i), rhs.get_unchecked(i))
         })
     }
-
     /// Extracts a slice containing the entire array.
     #[inline]
     pub fn as_slice(&self) -> &[T] {
@@ -241,6 +341,7 @@ where
     #[inline]
     pub fn from_slice(slice: &[T]) -> &GenericArray<T, N> {
         assert_eq!(slice.len(), N::to_usize());
+
         unsafe { &*(slice.as_ptr() as *const GenericArray<T, N>) }
     }
 
@@ -250,112 +351,8 @@ where
     #[inline]
     pub fn from_mut_slice(slice: &mut [T]) -> &mut GenericArray<T, N> {
         assert_eq!(slice.len(), N::to_usize());
+
         unsafe { &mut *(slice.as_mut_ptr() as *mut GenericArray<T, N>) }
-    }
-
-    /// Recursively converts one `GenericArray` to another via the `From` trait.
-    ///
-    /// If `From::from` panics, any already initialized elements will be dropped.
-    #[inline]
-    pub fn convert<U>(self) -> GenericArray<U, N>
-    where
-        U: From<T>,
-        N: ArrayLength<U>,
-    {
-        self.map(From::from)
-    }
-}
-
-impl<T, N> GenericArray<T, N>
-where
-    N: IterableArrayLength<T, N>,
-{
-    /// Recursively initializes a new `GenericArray` instance using the given function.
-    ///
-    /// If the generator function panics while initializing the array,
-    /// any already initialized elements will be dropped.
-    #[inline]
-    pub fn generate_recursive<F>(f: F) -> GenericArray<T, N>
-    where
-        F: Fn(usize) -> T,
-    {
-        RecursiveArrayBuilder::generate(f)
-    }
-
-    /// Recursively map a function over a  slice to a `GenericArray`.
-    ///
-    /// The length of the slice *must* be equal to the length of the array.
-    ///
-    /// If the mapping function panics, any already initialized elements will be dropped.
-    #[inline]
-    pub fn map_slice_recursive<S, F: Fn(&S) -> T>(s: &[S], f: F) -> GenericArray<T, N> {
-        assert_eq!(s.len(), N::to_usize());
-
-        RecursiveArrayBuilder::generate(|i| f(unsafe { s.get_unchecked(i) }))
-    }
-
-    /// Recursively maps a `GenericArray` to another `GenericArray`.
-    ///
-    /// If the mapping function panics, any already initialized elements in the new array
-    /// will be dropped, AND any unused elements in the source array will also be dropped.
-    #[inline]
-    pub fn map_recursive<U, F>(self, f: F) -> GenericArray<U, N>
-    where
-        F: Fn(T) -> U,
-        N: IterableArrayLength<U, N>,
-    {
-        RecursiveArrayBuilder::map(self, f)
-    }
-
-    /// Recursively maps a `GenericArray` to another `GenericArray` by reference.
-    ///
-    /// If the mapping function panics, any already initialized elements will be dropped.
-    #[inline]
-    pub fn map_ref_recursive<U, F>(&self, f: F) -> GenericArray<U, N>
-    where
-        F: Fn(&T) -> U,
-        N: IterableArrayLength<U, N>,
-    {
-        RecursiveArrayBuilder::map_ref(self, f)
-    }
-
-    /// Recursively ombines two `GenericArray` instances and iterates through both of them,
-    /// initializing a new `GenericArray` with the result of the zipped mapping function.
-    ///
-    /// If the mapping function panics, any already initialized elements in the new array
-    /// will be dropped, AND any unused elements in the source arrays will also be dropped.
-    #[inline]
-    pub fn zip_recursive<B, U, F>(self, rhs: GenericArray<B, N>, f: F) -> GenericArray<U, N>
-    where
-        F: Fn(T, B) -> U,
-        N: ArrayLength<B> + IterableArrayLength<U, N>,
-    {
-        RecursiveArrayBuilder::zip(self, rhs, f)
-    }
-
-    /// Recursively combines two `GenericArray` instances and iterates through both of them by reference,
-    /// initializing a new `GenericArray` with the result of the zipped mapping function.
-    ///
-    /// If the mapping function panics, any already initialized elements will be dropped.
-    #[inline]
-    pub fn zip_ref_recursive<B, U, F>(&self, rhs: &GenericArray<B, N>, f: F) -> GenericArray<U, N>
-    where
-        F: Fn(&T, &B) -> U,
-        N: ArrayLength<B> + IterableArrayLength<U, N>,
-    {
-        RecursiveArrayBuilder::zip_ref(self, rhs, f)
-    }
-
-    /// Recursively converts one `GenericArray` to another via the `From` trait.
-    ///
-    /// If `From::from` panics, any already initialized elements will be dropped.
-    #[inline]
-    pub fn convert_recursive<U>(self) -> GenericArray<U, N>
-    where
-        U: From<T>,
-        N: IterableArrayLength<U, N>,
-    {
-        self.map_recursive(From::from)
     }
 }
 
@@ -368,21 +365,95 @@ where
     /// Length of the slice must be equal to the length of the array
     #[inline]
     pub fn clone_from_slice(list: &[T]) -> GenericArray<T, N> {
-        GenericArray::map_slice(list, |x: &T| x.clone())
+        Self::from_exact_iter(list.iter().cloned()).expect(
+            "Slice must be the same length as the array",
+        )
     }
 }
 
-impl<T: Clone, N> GenericArray<T, N>
+impl<T, N> GenericArray<T, N>
 where
-    N: IterableArrayLength<T, N>,
+    N: ArrayLength<T>,
 {
-    /// Recursively Construct a `GenericArray` from a slice by cloning its content
-    ///
-    /// Length of the slice must be equal to the length of the array
-    #[inline]
-    pub fn clone_from_slice_recursive(list: &[T]) -> GenericArray<T, N> {
-        assert_eq!(list.len(), N::to_usize());
+    pub fn from_exact_iter<I>(iter: I) -> Option<Self>
+    where
+        I: IntoIterator<Item = T>,
+        <I as IntoIterator>::IntoIter: ExactSizeIterator,
+    {
+        let iter = iter.into_iter();
 
-        GenericArray::map_slice(list, |x: &T| x.clone())
+        if iter.len() == N::to_usize() {
+            let mut destination = ArrayBuilder::new();
+
+            for (dst, src) in destination.array.iter_mut().zip(iter.into_iter()) {
+                unsafe {
+                    ptr::write(dst, src);
+                }
+
+                destination.position += 1;
+            }
+
+            let array = unsafe { ptr::read(&destination.array) };
+
+            mem::forget(destination);
+
+            Some(ManuallyDrop::into_inner(array))
+        } else {
+            None
+        }
+    }
+}
+
+impl<T, N> ::core::iter::FromIterator<T> for GenericArray<T, N>
+where
+    N: ArrayLength<T>,
+    T: Default,
+{
+    fn from_iter<I>(iter: I) -> GenericArray<T, N>
+    where
+        I: IntoIterator<Item = T>,
+    {
+        let mut destination = ArrayBuilder::new();
+
+        let defaults = ::core::iter::repeat(()).map(|_| T::default());
+
+        for (dst, src) in destination.array.iter_mut().zip(
+            iter.into_iter().chain(defaults),
+        )
+        {
+            unsafe {
+                ptr::write(dst, src);
+            }
+        }
+
+        destination.into_inner()
+    }
+}
+
+#[cfg(test)]
+mod test {
+    // Compile with:
+    // cargo rustc --lib --profile test --release --
+    //      -C target-cpu=native -C opt-level=3 --emit asm
+    // and view the assembly to make sure test_assembly generates
+    // SIMD instructions instead of a niave loop.
+
+    #[inline(never)]
+    pub fn black_box<T>(val: T) -> T {
+        use core::{mem, ptr};
+
+        let ret = unsafe { ptr::read_volatile(&val) };
+        mem::forget(val);
+        ret
+    }
+
+    #[test]
+    fn test_assembly() {
+        let a = black_box(arr![i32; 1, 3, 5, 7]);
+        let b = black_box(arr![i32; 2, 4, 6, 8]);
+
+        let c = a.zip_ref(&b, |l, r| l + r);
+
+        assert_eq!(c, arr![i32; 3, 7, 11, 15]);
     }
 }
